For now ignore win/loss state entirely and focus on validity instead.
Try AI player against each possible game board on the final (9th) turn.
Check whether the suggested move is valid and update the counters
accordingly.

Validity is a simpler goal and a requirement to winning the game
anyway, so it makes sense to tackle this sub-problem first. The above
approach ensures that all parameter sets use the exact same test setup,
so the only random factor now are the parameters themselves.